requests: replaces status delegate with broadcast receiver
