+++ textile/src/androidTest/java/io/textile/textile/TextileTestListener.java
@@ -7,31 +7,31 @@ public class TextileTestListener extends BaseTextileEventListener {
     @Override
     public void nodeStarted() {
         super.nodeStarted();
-        System.out.println(":::::> node started");
+        System.out.println("::::::::::> node started");
     }
 
     @Override
     public void nodeStopped() {
         super.nodeStopped();
-        System.out.println(":::::> node stopped");
+        System.out.println("::::::::::> node stopped");
     }
 
     @Override
     public void nodeOnline() {
         super.nodeOnline();
-        System.out.println(":::::> node online");
+        System.out.println("::::::::::> node online");
     }
 
     @Override
     public void threadAdded(String threadId) {
         super.nodeStarted();
-        System.out.println(":::::> thread added: " + threadId);
+        System.out.println("::::::::::> thread added: " + threadId);
     }
 
     @Override
     public void threadRemoved(String threadId) {
         super.nodeStarted();
-        System.out.println(":::::> thread removed: " + threadId);
+        System.out.println("::::::::::> thread removed: " + threadId);
     }
 
     @Override
@@ -43,18 +43,18 @@ public class TextileTestListener extends BaseTextileEventListener {
         } else {
             progress = 0;
         }
-        System.out.println(":::::> sync update " + status.getId() + ": " + progress + "%");
+        System.out.println("::::::::::> sync update " + status.getId() + ": " + progress + "%");
     }
 
     @Override
     public void syncComplete(CafeSyncGroupStatus status) {
         super.nodeStarted();
-        System.out.println(":::::> sync complete: " + status.getId());
+        System.out.println("::::::::::> sync complete: " + status.getId());
     }
 
     @Override
     public void syncFailed(CafeSyncGroupStatus status) {
         super.nodeStarted();
-        System.out.println(":::::> sync failed: " + status.getId());
+        System.out.println("::::::::::> sync failed: " + status.getId());
     }
 }
