+++ textile/src/main/java/io/textile/textile/MessageHandler.java
@@ -1,10 +1,8 @@
 package io.textile.textile;
 
+import net.gotev.uploadservice.Logger;
+
 import java.util.Set;
-import java.util.logging.ConsoleHandler;
-import java.util.logging.Level;
-import java.util.logging.Logger;
-import java.util.logging.SimpleFormatter;
 
 import io.textile.pb.MessageOuterClass.Error;
 import io.textile.pb.Mobile.MobileEventType;
@@ -22,16 +20,9 @@ import mobile.Messenger;
 class MessageHandler implements Messenger {
 
     private Set<TextileEventListener> listeners;
-    private Logger logger;
 
     MessageHandler(final Set<TextileEventListener> listeners) {
         this.listeners = listeners;
-
-        this.logger = Logger.getLogger("textile.requests");
-        this.logger.setLevel(Level.ALL);
-        final ConsoleHandler handler = new ConsoleHandler();
-        handler.setFormatter(new SimpleFormatter());
-        this.logger.addHandler(handler);
     }
 
     @Override
@@ -80,7 +71,7 @@ class MessageHandler implements Messenger {
                             break;
                     }
                 } catch (final Exception e) {
-                    logger.warning(e.getMessage());
+                    Logger.error(getClass().getSimpleName(), e.getMessage());
                 }
                 break;
             case THREAD_UPDATE:
@@ -91,7 +82,7 @@ class MessageHandler implements Messenger {
                         listener.threadUpdateReceived(feedItem.getThread(), data);
                     }
                 } catch (final Exception e) {
-                    logger.warning(e.getMessage());
+                    Logger.error(getClass().getSimpleName(), e.getMessage());
                 }
                 break;
             case NOTIFICATION:
@@ -101,7 +92,7 @@ class MessageHandler implements Messenger {
                         listener.notificationReceived(notification);
                     }
                 } catch (final Exception e) {
-                    logger.warning(e.getMessage());
+                    Logger.error(getClass().getSimpleName(), e.getMessage());
                 }
                 break;
             case QUERY_RESPONSE:
@@ -136,7 +127,7 @@ class MessageHandler implements Messenger {
                             break;
                     }
                 } catch (final Exception e) {
-                    logger.warning(e.getMessage());
+                    Logger.error(getClass().getSimpleName(), e.getMessage());
                 }
                 break;
             case CAFE_SYNC_GROUP_UPDATE:
@@ -146,7 +137,7 @@ class MessageHandler implements Messenger {
                         listener.syncUpdate(status);
                     }
                 } catch (final Exception e) {
-                    logger.warning(e.getMessage());
+                    Logger.error(getClass().getSimpleName(), e.getMessage());
                 }
                 break;
             case CAFE_SYNC_GROUP_COMPLETE:
@@ -156,7 +147,7 @@ class MessageHandler implements Messenger {
                         listener.syncComplete(status);
                     }
                 } catch (final Exception e) {
-                    logger.warning(e.getMessage());
+                    Logger.error(getClass().getSimpleName(), e.getMessage());
                 }
                 break;
             case CAFE_SYNC_GROUP_FAILED:
@@ -166,7 +157,7 @@ class MessageHandler implements Messenger {
                         listener.syncFailed(status);
                     }
                 } catch (final Exception e) {
-                    logger.warning(e.getMessage());
+                    Logger.error(getClass().getSimpleName(), e.getMessage());
                 }
                 break;
         }
